requirements.txt
discord.py
gkeepapi
groq

python-dotenv
flask
